separate files IO file
