+++ specparam/objs/base.py
@@ -9,8 +9,8 @@ from specparam.core.utils import unlog
 from specparam.core.items import OBJ_DESC
 from specparam.core.errors import NoDataError
 from specparam.io.utils import get_files
-from specparam.io.models import (save_model, save_group, save_event,
-                                 load_json, load_jsonlines)
+from specparam.io.files import load_json, load_jsonlines
+from specparam.io.models import save_model, save_group, save_event
 from specparam.core.modutils import copy_doc_func_to_method
 from specparam.objs.results import BaseResults, BaseResults2D, BaseResults2DT, BaseResults3D
 from specparam.objs.data import BaseData, BaseData2D, BaseData2DT, BaseData3D
